Split out `DockerEngineApi`
